Working Commit:
Need to parallelize multiple back end calls using a fan out / fan in pattern.
Since there is a asyncio RPC client, I will consider adding an 0mq event
loop to the runtime. Another option would be to use a ThreadPool
executor. It really depends on the overhead and what plays nicest with
gevent. Using celery in this instance would be overkill, IMHO