MRC-110: 405 for /api/active/computers/query
**Story**
https://jira.rax.io/browse/MRC-110

**Summary**
Flask prioritizes /api/active/computers/<mercury_id> route
over /api/active/computers/query.
Hence, a "GET /api/active/computers/query" was matching the former and returning a 404.
However, it is an invalid method (only POST allowed on that endpoint) and should return a 405 instead.

This PR adds a converter that fails matching the former route by checking
if the mercury_id belongs to a set of blacklisted values (eg:query).
This forces the second route to be matched and a 405 to be returned subsequently.

Also, cleans up some of the old tests and adds relevant tests for the change.

**Testing**
tox